requirements.txt
langgraph
gradio
python-dotenv
langchain-community
together
openai
pandas
numpy
fastapi
uvicorn
python-multipart
openpyxl
tabula-py
pdfplumber
scikit-learn
langchain-google-genai
langchain-core
